futures_buffered
Struct FuturesOrdered

pub struct FuturesOrdered<T: Future> { /* private fields */ }

An unbounded queue of futures.

This โcombinatorโ is similar to FuturesUnordered, but it imposes an order on top of the set of futures. While futures in the set will race to completion in parallel, results will only be returned in the order their originating futures were added to the queue.

Futures are pushed into this queue and their realized values are yielded in order. This structure is optimized to manage a large number of futures. Futures managed by FuturesOrdered will only be polled when they generate notifications. This reduces the required amount of work needed to coordinate large numbers of futures.

When a FuturesOrdered is first created, it does not contain any futures. Calling poll in this state will result in Poll::Ready(None)) to be returned. Futures are submitted to the queue using push; however, the future will not be polled at this point. FuturesOrdered will only poll managed futures when FuturesOrdered::poll is called. As such, it is important to call poll after pushing new futures.

If FuturesOrdered::poll returns Poll::Ready(None) this means that the queue is currently not managing any futures. A future may be submitted to the queue at a later time. At that point, a call to FuturesOrdered::poll will either return the futureโs resolved value or Poll::Pending if the future has not yet completed. When multiple futures are submitted to the queue, FuturesOrdered::poll will return Poll::Pending until the first future completes, even if some of the later futures have already completed.

Note that you can create a ready-made FuturesOrdered via the collect method, or you can start with an empty queue with the FuturesOrdered::new constructor.

Implementations

impl<Fut: Future> FuturesOrdered<Fut>

pub fn new() -> Self

Constructs a new, empty FuturesOrdered

The returned FuturesOrdered does not contain any futures and, in this state, FuturesOrdered::poll_next will return Poll::Ready(None).

pub fn with_capacity(capacity: usize) -> Self

Constructs a new, empty FuturesOrdered

The returned FuturesOrdered does not contain any futures and, in this state, FuturesOrdered::poll_next will return Poll::Ready(None).

pub fn len(&self) -> usize

Returns the number of futures contained in the queue.

This represents the total number of in-flight futures, both those currently processing and those that have completed but which are waiting for earlier futures to complete.

pub fn is_empty(&self) -> bool

Returns true if the queue contains no futures

pub fn push_back(&mut self, future: Fut)

Pushes a future to the back of the queue.

This function submits the given future to the internal set for managing. This function will not call poll on the submitted future. The caller must ensure that FuturesOrderedBounded::poll is called in order to receive task notifications.

pub fn push_front(&mut self, future: Fut)

Pushes a future to the front of the queue.

This function submits the given future to the internal set for managing. This function will not call poll on the submitted future. The caller must ensure that FuturesOrderedBounded::poll is called in order to receive task notifications. This future will be the next future to be returned complete.