futures_buffered

Trait TryFuture

pub trait TryFuture: Future<Output = Result<Self::Ok, Self::Err>> + Sealed {
    type Ok;
    type Err;
}

A convenience for futures that return Result values that includes a variety of adapters tailored to such futures.

This is futures::TryFuture except it’s stricter on the future super-trait.

Required Associated Types

type Ok

type Err