futures_buffered

Trait TryStream

pub trait TryStream: Stream<Item = Result<Self::Ok, Self::Err>> + Sealed {
    type Ok;
    type Err;
}

A convenience for streams that return Result values that includes a variety of adapters tailored to such futures.

This is futures::TryStream except it’s stricter on the stream super-trait.

Required Associated Types

type Ok

type Err